chore: discovery tests simplificaiton
