+++ internal/discovery/concurrent_parse_race_test.go
@@ -2,7 +2,6 @@ package discovery_test
 
 import (
 	"fmt"
-	"os"
 	"path/filepath"
 	"strings"
 	"testing"
@@ -10,7 +9,6 @@ import (
 	"github.com/gruntwork-io/terragrunt/internal/component"
 	"github.com/gruntwork-io/terragrunt/internal/discovery"
 	"github.com/gruntwork-io/terragrunt/internal/filter"
-	"github.com/gruntwork-io/terragrunt/internal/venv"
 	"github.com/gruntwork-io/terragrunt/pkg/options"
 	"github.com/gruntwork-io/terragrunt/test/helpers"
 	"github.com/gruntwork-io/terragrunt/test/helpers/logger"
@@ -23,8 +21,10 @@ import (
 // at once, so one goroutine's parse stores the config while others read it.
 //
 // The shared unit gets a large config and many dependents, repeated across
-// several Discover calls, so the read/write overlap is reliably observable; a
-// smaller config or fewer iterations make the race intermittent.
+// several Discover calls, so the read/write overlap is reliably observable;
+// the config size and iteration count carry a margin over the smallest values
+// that still detect the race on every run. All fixture files live on an
+// in-memory filesystem, so only parsing costs wall time.
 //
 // To confirm the locks are load-bearing, drop the lock/unlock calls from Unit's
 // Config, StoreConfig, Reading, and SetReading and run with -race.
@@ -32,6 +32,8 @@ func TestDiscovery_GraphConcurrentConfigAccessWithRacing(t *testing.T) {
 	t.Parallel()
 
 	tmpDir := helpers.TmpDirWOSymlinks(t)
+
+	v := memGitTopLevelVenv(t, tmpDir)
 
 	// remote_state is partially decoded during discovery, so a large block is
 	// walked during parse rather than skipped, which lengthens the parse.
@@ -42,30 +44,27 @@ func TestDiscovery_GraphConcurrentConfigAccessWithRacing(t *testing.T) {
 		"  generate = { path = \"backend.tf\", if_exists = \"overwrite\" }\n  config = {\n",
 	)
 
-	for i := range 8000 {
+	for i := range 2000 {
 		fmt.Fprintf(&sharedConfig, "    k%d = \"v%d\"\n", i, i)
 	}
 
 	sharedConfig.WriteString("  }\n}\n")
 
-	vpcDir := filepath.Join(tmpDir, "vpc")
-	require.NoError(t, os.MkdirAll(vpcDir, 0755))
-	require.NoError(
-		t,
-		os.WriteFile(filepath.Join(vpcDir, "terragrunt.hcl"), []byte(sharedConfig.String()), 0644),
-	)
+	units := map[string]string{
+		filepath.Join(tmpDir, "vpc"): sharedConfig.String(),
+	}
 
 	const leaves = 8
 
 	for i := range leaves {
-		leafDir := filepath.Join(tmpDir, fmt.Sprintf("app%d", i))
-		require.NoError(t, os.MkdirAll(leafDir, 0755))
-		require.NoError(t, os.WriteFile(filepath.Join(leafDir, "terragrunt.hcl"), []byte(`
+		units[filepath.Join(tmpDir, fmt.Sprintf("app%d", i))] = `
 		dependency "vpc" {
 			config_path = "../vpc"
 		}
-		`), 0644))
+		`
 	}
+
+	writeUnits(t, v.FS, units)
 
 	l := logger.CreateLogger()
 	opts := &options.TerragruntOptions{
@@ -76,12 +75,12 @@ func TestDiscovery_GraphConcurrentConfigAccessWithRacing(t *testing.T) {
 	filters, err := filter.ParseFilterQueries(l, []string{"{./**}..."})
 	require.NoError(t, err)
 
-	for range 8 {
+	for range 4 {
 		d := discovery.NewDiscovery(tmpDir).
 			WithDiscoveryContext(&component.DiscoveryContext{WorkingDir: tmpDir}).
 			WithFilters(filters)
 
-		_, err := d.Discover(t.Context(), l, venv.OSVenv(), opts)
+		_, err := d.Discover(t.Context(), l, v, opts)
 		require.NoError(t, err)
 	}
 }
